backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.27.0
motor>=3.3.2
pydantic>=2.5.2
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
email-validator>=2.1.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
